ReadMe
This is a simple program that automatically unpacks every .dcx file in the directory. In order for it to work, you must download https://github.com/Atvaark/BinderTool and place in the same directory as the .dcx files and the UnpackDCX.jar. Run the JAR and it will automatically unpack every .dcx in the current directory.